+++ angular/CHANGELOG.md
@@ -27,7 +27,11 @@
     synthetic events (e.g. those triggered by a mouse click) instead of throwing
     a `TypeError`.
 
+*   [#1669][]: Fixed a regression which prevented a pipe invoked with more than
+    two arguments from being passed as an argument to a function.
+
 [#1694]: https://github.com/dart-lang/angular/issues/1694
+[#1669]: https://github.com/dart-lang/angular/issues/1669
 
 ## 5.2.0